Update QDecDouble.cc
It is necessary to use QString::number() instead of snprintf(str, MaxStrSize, "%.*g",...),
because snprintf() honors the application locale, which can lead to the fact,
that the radix character is not the point anymore. In this case fromString()
can not parse the string representation.

This behavior could be reproduced, if the locale LC_NUMERIC is de_DE for example.